Add script which runs the ETL script in the right environment
The configuration data needs to get munged in particular ways, and
then reset, and then the app hash needs to be recomputed in order for
the ETL script to work properly. It turned out to be easier to
script the whole process than to write docs explaining what needed
to happen.

Depends on https://github.com/oneiro-ndev/ndau/pull/89.