Add utility which writes the genesisfile to the chaos chain
This is designed to overwrite only values which are specifically
overridden; all other data already present on the chain remains there.

This genesis procedure will work before genesis or on a paused
blockchain instance, but it will change the root hash, so if the
chain height > 0, one should remember to update the hash that ndau
expects as well.